+++ remy/internal/injector/methods.go
@@ -2,6 +2,7 @@ package injector
 
 import (
 	"github.com/wrapped-owls/goremy-di/remy/internal/types"
+	"github.com/wrapped-owls/goremy-di/remy/pkg/keyopts"
 	"github.com/wrapped-owls/goremy-di/remy/pkg/utils"
 )
 
@@ -11,7 +12,7 @@ func Register[T any](ij types.Injector, bind types.Bind[T], keys ...string) erro
 		key = keys[0]
 	}
 
-	elementType := utils.GetKey[T](ij.ReflectOpts())
+	elementType := utils.GetKey[T](keyopts.FromReflectOpts(ij.ReflectOpts()))
 	var retriever types.DependencyRetriever = ij
 	if wrappedRetriever := retriever.WrapRetriever(); wrappedRetriever != nil {
 		retriever = wrappedRetriever
@@ -39,7 +40,7 @@ func Get[T any](retriever types.DependencyRetriever, keys ...string) (T, error) 
 	if len(keys) > 0 {
 		key = keys[0]
 	}
-	elementType := utils.GetKey[T](retriever.ReflectOpts())
+	elementType := utils.GetKey[T](keyopts.FromReflectOpts(retriever.ReflectOpts()))
 
 	var (
 		bind any
@@ -74,12 +75,21 @@ func TryGet[T any](retriever types.DependencyRetriever, keys ...string) (result 
 	return
 }
 
-func GetGen[T any](retriever types.DependencyRetriever, elements []types.InstancePair[any], keys ...string) (
-	result T, err error,
-) {
+func GetGen[T any](
+	retriever types.DependencyRetriever, elements []types.InstancePair[any], keys ...string,
+) (result T, err error) {
 	subInjector := New(false, retriever.ReflectOpts(), retriever)
 	for _, element := range elements {
-		bindKey := utils.GetElemKey(element.Value, subInjector.ReflectOpts())
+		var (
+			opts       = keyopts.FromReflectOpts(subInjector.ReflectOpts())
+			typeSeeker = element.Value
+		)
+		if element.InterfaceValue != nil {
+			opts |= keyopts.KeyOptIgnorePointer
+			typeSeeker = element.InterfaceValue
+		}
+		bindKey := utils.GetElemKey(typeSeeker, opts)
+
 		if element.Key != "" {
 			if err = subInjector.BindNamed(bindKey, element.Key, element.Value); err != nil {
 				return