+++ remy/internal/injector/methods_test.go
@@ -1,11 +1,14 @@
 package injector
 
 import (
+	"errors"
 	"fmt"
 	"testing"
 
 	"github.com/wrapped-owls/goremy-di/remy/internal/binds"
 	"github.com/wrapped-owls/goremy-di/remy/internal/types"
+	"github.com/wrapped-owls/goremy-di/remy/pkg/utils"
+	"github.com/wrapped-owls/goremy-di/remy/test/fixtures"
 )
 
 // TestGenerateBind__InstanceFactory verify if when registering an instance, it is only generated once
@@ -115,7 +118,11 @@ func TestRegister__Singleton(testObj *testing.T) {
 				for index := 0; index < 11; index++ {
 					_ = Register(i, sgtBind)
 					if invocations != bindCase.registerGenerations {
-						t.Errorf("Singleton %d times. Expected %d", invocations, bindCase.registerGenerations)
+						t.Errorf(
+							"Singleton %d times. Expected %d",
+							invocations,
+							bindCase.registerGenerations,
+						)
 						t.FailNow()
 					}
 				}
@@ -177,6 +184,7 @@ func TestRegister__overrideInstanceByBind(t *testing.T) {
 func TestGetGen(t *testing.T) {
 	const expected = "I love Go, yes this is true, as the answer 42"
 
+	interfaceValue := fixtures.GoProgrammingLang{}
 	testCases := [...]struct {
 		name           string
 		getGenCallback func(ij types.Injector) string
@@ -197,6 +205,10 @@ func TestGetGen(t *testing.T) {
 						{
 							Value: true,
 						},
+						{
+							Value:          interfaceValue,
+							InterfaceValue: (*fixtures.Language)(nil),
+						},
 					},
 				)
 			},
@@ -209,6 +221,10 @@ func TestGetGen(t *testing.T) {
 						err := Register(ij, binds.Instance[uint8](42))
 						err = Register(ij, binds.Instance("Go"), "lang")
 						err = Register(ij, binds.Instance(true))
+						err = Register[fixtures.Language](
+							ij,
+							binds.Instance[fixtures.Language](interfaceValue),
+						)
 						return err
 					},
 				)
@@ -220,11 +236,20 @@ func TestGetGen(t *testing.T) {
 		i := New(true, types.ReflectionOptions{})
 		_ = Register(
 			i, binds.Factory(
-				func(ij types.DependencyRetriever) (result string, err error) {
+				func(retriever types.DependencyRetriever) (result string, err error) {
 					result = fmt.Sprintf(
 						"I love %s, yes this is %v, as the answer %d",
-						TryGet[string](ij, "lang"), TryGet[bool](ij), TryGet[uint8](ij),
+						TryGet[string](
+							retriever,
+							"lang",
+						),
+						TryGet[bool](retriever),
+						TryGet[uint8](retriever),
 					)
+
+					if _, err = Get[fixtures.Language](retriever); err != nil {
+						t.Error(err)
+					}
 					return
 				},
 			),
@@ -240,20 +265,84 @@ func TestGetGen(t *testing.T) {
 				if result != expected {
 					t.Errorf(
 						"The direct params was not injected correctly.\nExpected: `%s`\nReceived: `%s`",
-						expected, result,
+						expected,
+						result,
 					)
 					t.FailNow()
 				}
 
 				// Check if the binds doesn't exist after do the GetGen
-				uintResult := TryGet[uint8](i)
-				boolResult := TryGet[bool](i)
-				strResult := TryGet[string](i, "lang")
-
+				var (
+					uintResult = TryGet[uint8](i)
+					boolResult = TryGet[bool](i)
+					strResult  = TryGet[string](i, "lang")
+				)
 				if uintResult != 0 || boolResult || len(strResult) > 0 {
 					t.Error("Parameter injection values override the original injector")
 				}
 			},
 		)
 	}
+}
+
+func TestGetGen_raiseCastError(t *testing.T) {
+	var (
+		i                                = New(true, types.ReflectionOptions{})
+		interfaceValue fixtures.Language = fixtures.GoProgrammingLang{}
+	)
+	err := Register(
+		i, binds.Factory(
+			func(retriever types.DependencyRetriever) (result string, getErr error) {
+				var lang fixtures.Language
+				if lang, getErr = Get[fixtures.Language](retriever); getErr == nil {
+					result = lang.Kind() + " language: " + lang.Name()
+				}
+				return
+			},
+		),
+	)
+	if err != nil {
+		t.Error(err)
+		t.FailNow()
+	}
+
+	t.Run(
+		"Correctly bind registration", func(t *testing.T) {
+			_, err = GetGen[string](
+				i,
+				[]types.InstancePair[any]{
+					{
+						Value:          interfaceValue,
+						InterfaceValue: (*fixtures.Language)(nil),
+					},
+				},
+			)
+			if err != nil {
+				t.Error(err)
+				t.FailNow()
+			}
+		},
+	)
+
+	t.Run(
+		"Register pointer interface value", func(t *testing.T) {
+			_, err = GetGen[string](
+				i,
+				[]types.InstancePair[any]{
+					{
+						Value:          &interfaceValue,
+						InterfaceValue: (*fixtures.Language)(nil),
+					},
+				},
+			)
+			if err == nil {
+				t.Error("No error has returned after binding the value incorrectly")
+				t.FailNow()
+			}
+
+			if !errors.Is(err, utils.ErrTypeCastInRuntime) {
+				t.Errorf("Unknown error raised: `%v`\n", err)
+			}
+		},
+	)
 }